# Optimize HMM implementation